src/authorization-service/requirements.txt
aws-cdk-lib==2.148.0
constructs>=10.0.0,<11.0.0
python-dotenv
